# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.accountResp:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_activated:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
  api.authResp:
    properties:
      token:
        type: string
      type:
        type: string
    type: object
  api.defaultJsonResp:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  api.loginReq:
    properties:
      password:
        type: string
      phone_number:
        type: string
    required:
    - password
    - phone_number
    type: object
  api.registerReq:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      password_confirmation:
        type: string
      phone_number:
        type: string
    required:
    - first_name
    - last_name
    - password
    - password_confirmation
    - phone_number
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Deincom Apis
  title: Deeincom APIs
  version: "1.0"
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: register an account
      parameters:
      - description: Register account information
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.registerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.accountResp'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.defaultJsonResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.defaultJsonResp'
      summary: register
      tags:
      - accounts
  /accounts/auth:
    post:
      consumes:
      - application/json
      description: get token using credentials
      parameters:
      - description: account credentials
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResp'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.defaultJsonResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.defaultJsonResp'
      summary: authenticate
      tags:
      - accounts
  /accounts/profile:
    get:
      consumes:
      - application/json
      description: get account profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.accountResp'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.defaultJsonResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.defaultJsonResp'
      security:
      - ApiBearerKey: []
      summary: account profiles
      tags:
      - accounts
  /accounts/verification:
    get:
      consumes:
      - application/json
      description: sent sms verify code
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.defaultJsonResp'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.defaultJsonResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.defaultJsonResp'
      security:
      - ApiBearerKey: []
      summary: sent verify code
      tags:
      - accounts
  /accounts/verification/{code}/code:
    get:
      consumes:
      - application/json
      description: sent sms verify code
      parameters:
      - description: verify code
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResp'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.defaultJsonResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.defaultJsonResp'
      security:
      - ApiBearerKey: []
      summary: sent verify code
      tags:
      - accounts
schemes:
- http
- https
securityDefinitions:
  ApiBearerKey:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"